supposedly fixes failing juju-k8s test
juju has background task performing status synchs and checking
for snap packages updates periodically. a side effect of their
activities is a restart of kubelet service. this often happens
when the test is in progress and doesn't expect a collapse of its
infrastructure.
let's tell juju to perform the jobs less often, use dedicated
config options.

Change-Id: I1e1121d5687a000029bca96ffeb22302a2eb9e7a