sync time
do the same as for other deployers i.e. sync ntp at all target nodes.
unfortunatelly it's imposible to get the nodes list from oc because
they are not there almost until the end of tf stage thus it is
reasonable to obtain the nodes IPs from the DHCP leases of the
openshift network in libvirt, use virsh and then call standart sync_time

Change-Id: Ife810a9ef671cb62230f8efdbc2c673cf3d27958
Signed-off-by: sgalas@progmaticlab.com